feat(guilds): complete the world-facing half of the guild system
The social half (create/edit/join/manage/forums) was real; everything that
made a guild visible in-world was missing, and four header constants marked
"AS3-verified" were wrong.

Live bugs fixed
- Guild rooms stripped build rights: CanManipulateFurniAsync required
  GroupAdmin, a level nothing produced, so classic rights-holders lost build
  rights the moment a guild was created and members gained none.
- KickMemberMessageEvent was 781, which is really GetMemberGuildItemCount, so
  kicking fired on the furni-count request. Kick is 3156.
- HabboGroupDeactivatedMessageComposer was 2087, which is really
  GroupMembershipRequested, so the client over-read the buffer on every guild
  dissolution. Deactivated is 1948.
- Guild navigator searches returned the whole hotel (NavigatorQueryType.MyGroups
  had no case); "group:" search filtered room names instead of guild names.
- Guild creation could burn credits: the unique index on groups.room_id blocked
  reuse after a soft delete, and the debit had no refund path.

Added
- Guild room rights: RoomGroupContext + RoomSecurityPolicy resolve
  GroupRights/GroupAdmin from a roster hydrated into RoomLiveState and kept live
  by GroupGrain -> RoomGrain.RefreshGroupMembershipAsync.
- Room mod settings (WhoCanMute/Kick/Ban) are now enforced at all — they were
  persisted and echoed but checked nowhere, so any occupant could mute/kick/ban.
- Favourite-guild badge on avatars, pushed live via FavoriteMembershipUpdate;
  IsGroupMember on room cards; guild list on extended profiles.
- Forum read markers and real unread counts; listCode sorting (most active /
  most viewed / my forums) with a real view counter; moderator attribution.
- Guild block list (kick-with-block + unblock), admin notification of pending
  join requests, guild name validation on rename.
- Guild furni: stuff-data stamped at purchase, context menu answered,
  furniture_guild_customized logic bound.
- Badge/colour palettes now actually seeded (the .sql files were referenced by
  nothing, leaving the badge editor empty on a fresh DB).

groups.room_id is no longer unique (EF's relationship convention forced it; the
[Index] attribute cannot undo that, hence the fluent override). MySQL has no
partial indexes, so "one live guild per room" moves to nullable rooms.group_id.

Quality gate green; 334 tests pass, including 14 new permission cases.

+++ Vortex.Revisions/Revision20260701/Serializers/Users/HabboGroupDetailsMessageComposerSerializer.cs
@@ -22,7 +22,7 @@ internal class HabboGroupDetailsMessageComposerSerializer(int header)
         packet.WriteString(details.BadgeCode);
         packet.WriteInteger(details.RoomId);
         packet.WriteString(details.RoomName);
-        packet.WriteInteger(details.Status);
+        packet.WriteInteger((int)details.Status);
         packet.WriteInteger(details.TotalMembers);
         packet.WriteBoolean(details.Favourite);
         packet.WriteString(details.CreationDate);
